backend/requirements.txt
# Backend Dependencies
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.11.2

# API Documentation
drf-yasg==1.21.7

# HTTP Requests
requests==2.31.0

# ML & Data
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
